+++ hamcrest-matchers/java7-hamcrest-matchers/src/main/java/it/ozimov/cirneco/hamcrest/java7/clazz/IsValidNoArgumentConstructor.java
@@ -13,7 +13,7 @@ import java.lang.reflect.Constructor;
  */
 public class IsValidNoArgumentConstructor extends TypeSafeMatcher<Class> {
 
-    public IsValidNoArgumentConstructor() {
+    private IsValidNoArgumentConstructor() {
     }
 
     /**