+++ hamcrest-matchers/java7-hamcrest-matchers/src/test/java/it/ozimov/cirneco/hamcrest/java7/clazz/TestClassWithLombokDataAnnotation.java
@@ -1,0 +1,13 @@
+package it.ozimov.cirneco.hamcrest.java7.clazz;
+
+import lombok.Data;
+
+
+@Data
+public class TestClassWithLombokDataAnnotation {
+
+    private TestClassWithLombokDataAnnotation() {
+
+    }
+
+}