+++ hamcrest-matchers/java7-hamcrest-matchers/src/test/java/it/ozimov/cirneco/hamcrest/java7/clazz/TestClassWithNoArgsConstructor.java
@@ -1,0 +1,8 @@
+package it.ozimov.cirneco.hamcrest.java7.clazz;
+
+public class TestClassWithNoArgsConstructor {
+
+    public TestClassWithNoArgsConstructor(String param) {
+
+    }
+}