+++ hamcrest-matchers/java7-hamcrest-matchers/src/test/java/it/ozimov/cirneco/hamcrest/java7/clazz/TestClassWithNoNonnullAnnotation.java
@@ -1,0 +1,9 @@
+package it.ozimov.cirneco.hamcrest.java7.clazz;
+
+public class TestClassWithNoNonnullAnnotation {
+
+    private TestClassWithNoNonnullAnnotation() {
+
+    }
+
+}