+++ hamcrest-matchers/java7-hamcrest-matchers/src/test/java/it/ozimov/cirneco/hamcrest/java7/clazz/TestClassWithNonnullAnnotation.java
@@ -1,0 +1,12 @@
+package it.ozimov.cirneco.hamcrest.java7.clazz;
+
+import javax.annotation.Nonnull;
+
+@Nonnull
+public class TestClassWithNonnullAnnotation {
+
+    private TestClassWithNonnullAnnotation() {
+
+    }
+
+}